requirements.txt
Flask
requests
Pillow
fpdf2
python-dotenv
gunicorn
